remove debug prints and comments, better error msg on failed submission, js sets utils
