+++ static/js/components/TransferList.jsx
@@ -11,22 +11,7 @@ import Checkbox from "@mui/material/Checkbox";
 import Button from "@mui/material/Button";
 import Divider from "@mui/material/Divider";
 
-function not(a, b) {
-  // we can't do indexOf on objects, so we need to compare the id
-  return a.filter(
-    (value) => b.findIndex((item) => item.id === value.id) === -1,
-  );
-}
-
-function intersection(a, b) {
-  return a.filter(
-    (value) => b.findIndex((item) => item.id === value.id) !== -1,
-  );
-}
-
-function union(a, b) {
-  return [...a, ...not(b, a)];
-}
+import { not, intersection, union } from "../utils";
 
 const TransferList = ({
   left,
